Preliminary refactor of /MassSpecMain/. d+c.
all the work for the whole class is done inside of the constructor,
and the organization within the frame is not ideal. further refactoring
is needed to section off the construction of the left panel into a
separate method which returns lP as a whole. same with tandem and output
graphs.